PROJECT TITLE: EDA - AMAZON BEST SELLER
Introduction
This research is based on the record of Amazon E-commerce data of one month. The research is conducted to find out the Best Amazon Seller of 2021 June. The growth of E-commerce is increasing and market competitions are also high in this scenario, According Amazon wants to find out the best sellers to improve the productivity of the organization. Here I am going to explore the data of Amazon's one month that is for 2021 June.

Attribute Information
Amazon_Best_Seller_2021_June is a dataset on Amazon's top 100 selling products 2021. Data has been categorized into:

1)ASIN
2)Category
3)Product Link
4)No of Sellers
5)Rating
6)Reviews Count
7)Price

About ASIN:

An Amazon Standard Identification Number (ASIN) is a 10-character alphanumeric unique identifier assigned by Amazon.com and its partners for product identification within the Amazon organization.

Type of Categories are as follow:

Electronics
Clothing, Shoes & Jewelry
Gift Cards
Books
Video Games
Camera & Photo

Conclusion
From the study we can convey only the overall performance of the sellers, we can't take individually, Because of the some limitations in the data. The dataset is incomplete with limited number of columns and rows.

Intertesting factor is that the sellers with Electonics sold highest amount of items. Amazon can consider sellers with electronics providing the best productivity for the organization.

Also Amazon have to boost the performance of other sellers for better productivity and goodwillof the organization.

Here we can conclude that the rating is best at its optimum level of 4.8 with the highest position.

In the case of reviews count we can consider the best in average value of 77005.94.

Must have to consider an ideal pricing.

The rating has mostly happened in 4.8 with the best rating in the category of Gift cards and the category with Gift cards has more reviews count compared to others.

Hence Amazon must have to focus on other categories also.

Not too much correlation in the dataset. The highest correlation for rating is reviews count. The correlation for no of sellers is category and rating and there are slight correlation for price with no of sellers and rank. Hence, we can see in the heatmap there is no ideal correlation.